Rename packages according to the directory name of their source files
 - Otherwise this would require importing the package with one name, but using it with another
(for example importing "github.com/philippgille/x/y", but calling a function with "z.SomeFunc()")
