+++ redis/redis.go
@@ -1,4 +1,4 @@
-package gokv
+package redis
 
 import (
 	"github.com/go-redis/redis"
